add new line at the end when rstrip

+++ rstrip.py
@@ -2,18 +2,24 @@
 """
 rstrip every line of file(s)
 """
-import sys
 
 
-def rstrip_file(fname):
+def rstrip_file(fname, newlines=1):
     with open(fname, 'r+') as f:
-        ss = f.readlines()
+        s = f.read()
+        if not s:
+            raise Exception('Empty file.')
+        ss = [line.rstrip() for line in s.rstrip().split('\n')]
+        required = '\n'.join(ss) + '\n'*newlines
+        if s == required:
+            raise Exception('Already meet requirement.')
         f.seek(0)
         f.truncate()
-        f.write('\n'.join([i.rstrip() for i in ss]))
+        f.write(required)
 
 
 def main():
+    import sys
     if not sys.argv[1:]:
         print(__doc__.strip())
         print('\nUsage:')
@@ -24,7 +30,7 @@ def main():
         try:
             rstrip_file(fn)
         except Exception as e:
-            print('{}: skip!'.format(fn))
+            print('{}: skip! {}'.format(fn, e))
         else:
             print('{}: rstriped.'.format(fn))
     print('Done!')
